id save at cookie and update, delete check cookie

+++ backend/routes/boardControl.js
@@ -1,13 +1,16 @@
 var express = require('express');
 var router = express.Router();
+var cookieParser = require('cookie-parser');
 const models = require('../models');
 let jwt = require('jsonwebtoken');
 let secretObj = require('../config/jwt');
 let verify = require('../config/verify');
 
+router.use(cookieParser('asd123'));
+
 router.get('/userview', function(req, res, next) {
 
-    let usertoken = req.headers.token;
+    let usertoken = req.signedCookies.token;
 
     if(verify(usertoken, secretObj.secret)) {
         models.user.findAll()
@@ -59,20 +62,40 @@ router.post('/addpost', function(req, res, next) {
 });
 
 router.delete('/delpost', function(req, res, next) {
+
+    
     
     let usertoken = req.headers.token;
+    // console.log(usertoken);
+    let id = req.signedCookies.userid;
+    // console.log(id);
+    
     
     let body = req.body;
 
     if(verify(usertoken, secretObj.secret)) {
-        models.post.destroy({
-            where : { no : body.no }
+
+        models.post.findOne({
+            where: { no : body.no}
         })
-        .then( result => {
-            res.json({ success : result });
+        .then( findpost => {
+            // console.log("찾은 정보 :", findpost.dataValues.writer);
+            if (findpost.dataValues.writer === id){
+                models.post.destroy({
+                    where : {no : findpost.dataValues.no}
+                })
+                .then ( del => {
+                    res.json({ success : del });
+                })
+                .catch ( err => {
+                    console.log(err);
+                })
+            }
+            else (console.log("권한이 없습니다."));
         })
         .catch( err => {
-            console.log("게시글 삭제에 실패했습니다")
+            console.log(err);
+            console.log("게시글 정보를 찾지 못했습니다.")
         })
     }
     else {
@@ -80,6 +103,50 @@ router.delete('/delpost', function(req, res, next) {
     }
 })
 
+router.put('/uppost', function(req, res, next) {
+    let usertoken = req.headers.token;
+    console.log(usertoken);
+    let id = req.signedCookies.userid;
+    console.log(id);
+
+    let body = req.body;
+
+    console.log(body);
+
+    if(verify(usertoken, secretObj.secret)) {
+        models.post.findOne({
+            where: { no : body.no}
+        })
+        .then( findpost => {
+            console.log("찾은 정보 :", findpost);
+            if (findpost.dataValues.writer === id){
+                models.post.update({
+                    title : body.title,
+                    content : body.content
+                    },
+                    {
+                        where : {no : findpost.dataValues.no}
+                })
+                .then ( updat => {
+                    console.log("수정 완료했습니다")
+                    res.json({ success : updat });
+                })
+                .catch ( err => {
+                    console.log(err);
+                    console.log("수정에 실패했습니다")
+                })
+            }
+            else (console.log("권한이 없습니다."));
+        })
+        .catch( err => {
+            console.log(err);
+            console.log("게시글 정보를 찾지 못했습니다.")
+        })
+    }
+    else {
+        res.json({fail})
+    }
+})
 
 router.post('/addreply', function(req, res, next) {
